[common] Add more fallback shells
The dpkg `control` specifies `zsh` and `fish` in addition to `bash` as
dependencies, but the app didn't use them directly.

This commit updates the app to attempt to use those shells when
`userPasswd` is unavailable, which somewhat seems to be the case in my
environment.

This also adds `sh` as an optional dependency, which may be optimum for
minimalistic setups.